Percent Change Function
This program will check the change in an account balance. Write a function that accepts a start_price and an end_price and returns the percent change using the relationship. Round the final result to the nearest whole number. Print a statement above the return that says: 'The percent change in the account balance is:'