Insurance Policy
Suppose your apartment was burglarized.  The insurance guy calls to discuss your policy, which is limited to cover a particular maximum in losses (for example, $50,000). You calculate the bill for the difference you lost.

Given an dictionary of the stolen items and a limit, return the difference between the total value of those items and the limit of the policy. Name the function calc_diff().